# src/doc/swagger.yaml
openapi: 3.0.4
info:
  title: Blog API
  version: 2.0.0
  description: "API documentation for personal blogging platform"
servers:
  - url: http://localhost:5000/api
    description: Local development server
...
paths:
...
  /users:
    get:
      security:
        - JWTAuth: []
      operationId: getUsers
      summary: Get a list of users
      description: Get a paginated list of users.  If the access token was refreshed during authentication, a new token will also be included.
      tags: [Users]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 10
          description: Max number of users to return

        - $ref: "#/components/parameters/CursorParam"

        - $ref: "#/components/parameters/sortParam"

        - in: query
          name: role
          schema:
            type: string
            enum: [admin, moderator, user]
          description: Filter users by role
          required: true

      responses:
        200:
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: Returned only if a new access token was generated
                    nullable: true
                  cursor:
                    type: string
                    description: Cursor for pagination
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: Unique identifier for the user
                        fullName:
                          type: string
                          description: Full name of the user
                        username:
                          type: string
                          description: Unique username of the user
                        profileImage:
                          type: string
                          description: URL of the user's profile image
              examples:
                withAccessToken:
                  summary: Example with access token
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR....
                    cursor: 66e038559b9bb0ffc46836cf
                    users:
                      - fullName: "John Doe"
                        username: "johndoe"
                        profileImage: "https://cdn.example.com/images/johndoe.png"
                      - fullName: "Jane Smith"
                        username: "janesmith"
                        profileImage: "https://cdn.example.com/images/janesmith.png"
                withoutAccessToken:
                  summary: Example without access token
                  value:
                    cursor: 66e038559b9bb0ffc46836cf
                    users:
                      - fullName: "John Doe"
                        username: "johndoe"
                        profileImage: "https://cdn.example.com/images/johndoe.png"
                      - fullName: "Jane Smith"
                        username: "janesmith"
                        profileImage: "https://cdn.example.com/images/janesmith.png"
        400:
          description: Bad Request, Invalid query parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Error message for invalid query parameter
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Use the JWT token obtained from login or registration to access protected routes.
  parameters:
...
    CursorParam:
      in: query
      name: cursor
      schema:
        type: string
        nullable: true
        example: 64f0341f3df05e12a3c3a401
      description: Cursor to the next page (null if no more)
    sortParam:
      in: query
      name: sort
      schema:
        type: string
        enum: [latest, oldest, top]
        default: latest
      description: Sort order
  responses:
...
    Forbidden:
      description: Forbidden – user lacks permission (read description to know what roles are authorized)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Forbidden
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error message
...